+++ torchkge/models/translation.py
@@ -427,7 +427,7 @@ class TransRModel(TranslationModel):
             candidates = self.rel_emb.weight.data.view(1, self.n_rel, self.rel_emb_dim)
             candidates = candidates.expand(b_size, self.n_rel, self.rel_emb_dim)
 
-        return proj_h, proj_t, candidates, r
+        return proj_h, proj_t, r, candidates
 
     def evaluate_projectionss(self):
         """Link prediction evaluation helper function. Project all entities
@@ -624,7 +624,7 @@ class TransDModel(TranslationModel):
             candidates = self.rel_emb.weight.data.view(1, self.n_rel, self.rel_emb_dim)
             candidates = candidates.expand(b_size, self.n_rel, self.rel_emb_dim)
 
-        return proj_h, proj_t, candidates, r
+        return proj_h, proj_t, r, candidates
 
     def evaluate_projectionss(self):
         """Link prediction evaluation helper function. Project all entities