+++ torchkge/utils/__init__.py
@@ -14,4 +14,5 @@ from .dissimilarities import l1_torus_dissimilarity, l2_torus_dissimilarity, \
 from .losses import MarginLoss, LogisticLoss, BinaryCrossEntropyLoss
 from .modeling import init_embedding, get_true_targets, load_embeddings, filter_scores
 from .operations import get_rank, get_mask, get_bernoulli_probs
+from .pretrained_models import load_pretrained_transe, load_pretrained_rescal, load_pretrained_complex
 from .training import Trainer, TrainDataLoader